+++ Crud/src/main/java/com/KnowIt/Gym_intellect_Crud/Repository/Workout_planrepo.java
@@ -4,7 +4,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.KnowIt.Gym_intellect_Crud.Entity.Workout_plan;
 
-public interface Workout_planrepo extends JpaRepository<Workout_plan,Long>{
+public interface Workout_planrepo extends JpaRepository<Workout_plan,Integer>{
 	List<Workout_plan> findByPlanName(String planName);
-	
 }
